+++ Sistema/Smartgym/Smartgym/Models/Professor.cs
@@ -59,9 +59,9 @@ namespace Smartgym.Models
 
         public DateTime DataAdmissaoProfessor { get; set; }
 
-        public int TelefoneProfessor { get; set; }
+        public long TelefoneProfessor { get; set; }
 
-        public int CelularProfessor { get; set; }
+        public long CelularProfessor { get; set; }
 
         [MinLength(1)]
         [MaxLength(1)]
